First Gtk and Qt refactoring
